testool: support test with subcircuit
