+++ zkevm-circuits/src/modexp_circuit.rs
@@ -1,6 +1,11 @@
 //! The Modexp circuit is responsible for modexp operations on big integer from precompiled contract
 //! calls ModExp, current the size of supported integer is up to 32 bytes (U256)
 
+#[cfg(any(feature = "test", test, feature = "test-circuits"))]
+mod dev;
+#[cfg(any(feature = "test", test, feature = "test-circuits"))]
+mod test;
+
 use halo2_proofs::{
     circuit::{Layouter, Region, Value},
     plonk::{Advice, Column, ConstraintSystem, Error},
@@ -196,94 +201,5 @@ impl<F: Field> SubCircuit<F> for ModExpCircuit<F> {
         )?;
 
         config.modexp_table.fill_blank(layouter)
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::util::MockChallenges;
-    use halo2_proofs::{
-        circuit::SimpleFloorPlanner,
-        dev::MockProver,
-        halo2curves::bn256::Fr,
-        plonk::{Circuit, ConstraintSystem},
-    };
-
-    impl Circuit<Fr> for ModExpCircuit<Fr> {
-        type Config = (ModExpCircuitConfig, MockChallenges);
-        type FloorPlanner = SimpleFloorPlanner;
-
-        fn without_witnesses(&self) -> Self {
-            Self::default()
-        }
-
-        fn configure(meta: &mut ConstraintSystem<Fr>) -> Self::Config {
-            let modexp_table = ModExpTable::construct(meta);
-            let challenge = MockChallenges::construct(meta);
-            (
-                <ModExpCircuitConfig as SubCircuitConfig<Fr>>::new(meta, modexp_table),
-                challenge,
-            )
-        }
-
-        fn synthesize(
-            &self,
-            (config, challenge): Self::Config,
-            mut layouter: impl Layouter<Fr>,
-        ) -> Result<(), Error> {
-            let challenges = challenge.values(&layouter);
-            <Self as SubCircuit<Fr>>::synthesize_sub(self, &config, &challenges, &mut layouter)
-        }
-    }
-
-    #[test]
-    fn test_modexp_circuit_00() {
-        let base = Word::from(1u128);
-        let exp = Word::from(2u128);
-        let modulus = Word::from(7u128);
-        let (_, result) = base.pow(exp).div_mod(modulus);
-        let event1 = BigModExp {
-            base,
-            exponent: exp,
-            modulus,
-            result,
-        };
-        let test_circuit = ModExpCircuit(vec![event1], Default::default());
-        let prover = MockProver::run(16, &test_circuit, vec![]).unwrap();
-        assert_eq!(prover.verify(), Ok(()));
-    }
-
-    #[test]
-    fn test_modexp_circuit_01() {
-        let base = Word::from(1u128);
-        let exp = Word::from(2u128);
-        let modulus = Word::from(7u128);
-        let (_, result) = base.pow(exp).div_mod(modulus);
-        let event1 = BigModExp {
-            base,
-            exponent: exp,
-            modulus,
-            result,
-        };
-        let test_circuit = ModExpCircuit(vec![event1], Default::default());
-        let prover = MockProver::run(16, &test_circuit, vec![]).unwrap();
-        assert_eq!(prover.verify(), Ok(()));
-    }
-    #[test]
-    fn test_modexp_circuit_02() {
-        let base = Word::from(2u128);
-        let exp = Word::from(2u128);
-        let modulus = Word::from(7u128);
-        let (_, result) = base.pow(exp).div_mod(modulus);
-        let event1 = BigModExp {
-            base,
-            exponent: exp,
-            modulus,
-            result,
-        };
-        let test_circuit = ModExpCircuit(vec![event1], Default::default());
-        let prover = MockProver::run(16, &test_circuit, vec![]).unwrap();
-        assert_eq!(prover.verify(), Ok(()));
     }
 }
